+++ index.php
@@ -32,6 +32,8 @@
 
 		<?php else : ?>
 
+			<?php do_action( 'inis_b_before_empty_archive' ); ?>
+
 			<article class="post not-found error-404">
 				<header>
 					<h2 class="post-title"><?php _e( 'Nothing Found', 'inis-b' ); ?></h2>
@@ -41,6 +43,8 @@
 					<p><?php _e('Sorry, but no posts were found.', 'inis-b'); ?></p>
 				</div>
 			</article>
+
+			<?php do_action( 'inis_b_after_empty_archive' ); ?>
 
 		<?php endif; ?>
 	</div>